379 connect vsi internal ip input (#390)
Expanded supported src/dst inputs to include vsi name and internal
